Updating s3cmd and fixing css/js mimetypes
By default, s3cmd will use python's "magic" module (which is a wrapper around
libmagic) to guess mimetypes when --guess-mime-type is given. This is
unfortunate because libmagic will identify .css and .js files as text/plain...
which can bork the site.

So we provide --no-mime-magic which forces a fallback to the python mimetypes
module which does the correct thing.